chore: Move middleware
