+++ api.planx.uk/modules/pay/service/inviteToPay/paymentRequest.ts
@@ -1,120 +1,9 @@
 import { gql } from "graphql-request";
-import { NextFunction, Request, Response } from "express";
-import { ServerError } from "../../../../errors";
+import { Request } from "express";
 import { fetchPaymentViaProxyWithCallback } from "../../controller";
 import { GovUKPayment } from "@opensystemslab/planx-core/types";
 import { $api } from "../../../../client";
 import { postPaymentNotificationToSlack } from "../utils";
-
-// https://docs.payments.service.gov.uk/api_reference/create_a_payment_reference/#json-body-parameters-for-39-create-a-payment-39
-interface GovPayCreatePayment {
-  amount: number;
-  reference: string;
-  description: string;
-  return_url: string;
-}
-
-interface GetPaymentRequestDetails {
-  paymentRequest: {
-    sessionId: string;
-    paymentAmount: number;
-    session: {
-      flowId: string;
-      flow: {
-        team: {
-          slug: string;
-        };
-      };
-    };
-  } | null;
-}
-
-// middleware used by routes:
-//  * /payment-request/:paymentRequest/pay
-//  * /payment-request/:paymentRequest/payment/:paymentId
-export async function fetchPaymentRequestDetails(
-  req: Request,
-  _res: Response,
-  next: NextFunction,
-) {
-  const query = gql`
-    query GetPaymentRequestDetails($paymentRequestId: uuid!) {
-      paymentRequest: payment_requests_by_pk(id: $paymentRequestId) {
-        sessionId: session_id
-        paymentAmount: payment_amount
-        session {
-          flowId: flow_id
-          flow {
-            team {
-              slug
-            }
-          }
-        }
-      }
-    }
-  `;
-  const { paymentRequest } =
-    await $api.client.request<GetPaymentRequestDetails>(query, {
-      paymentRequestId: req.params.paymentRequest,
-    });
-  if (!paymentRequest) {
-    return next(
-      new ServerError({
-        message: "payment request not found",
-        status: 404,
-      }),
-    );
-  }
-  const sessionId = paymentRequest.sessionId;
-  if (sessionId) req.query.sessionId = sessionId;
-
-  const localAuthority = paymentRequest.session?.flow?.team?.slug;
-  if (localAuthority) req.params.localAuthority = localAuthority;
-
-  const flowId = paymentRequest.session?.flowId;
-  if (flowId) req.query.flowId = flowId;
-
-  const paymentAmount = paymentRequest.paymentAmount.toString();
-  if (paymentAmount) req.params.paymentAmount = paymentAmount;
-
-  next();
-}
-
-// middleware used by /payment-request/:paymentRequest/pay?returnURL=...
-export async function buildPaymentPayload(
-  req: Request,
-  _res: Response,
-  next: NextFunction,
-) {
-  if (!req.query.returnURL) {
-    return next(
-      new ServerError({
-        message: "Missing required returnURL query param",
-        status: 400,
-      }),
-    );
-  }
-
-  if (!req.query.sessionId) {
-    return next(
-      new ServerError({
-        message: "Missing required sessionId query param",
-        status: 400,
-      }),
-    );
-  }
-
-  const createPaymentBody: GovPayCreatePayment = {
-    amount: parseInt(req.params.paymentAmount),
-    reference: req.query.sessionId as string,
-    description: "New application (nominated payee)",
-    return_url: req.query.returnURL as string,
-  };
-
-  req.body = createPaymentBody;
-
-  next();
-}
 
 export const fetchPaymentRequestViaProxy = fetchPaymentViaProxyWithCallback(
   async (req: Request, govUkResponse: GovUKPayment) => {
